<rdar://problem/7424387> WebCore binary missing symbols when built for x86_64 from a machine that cannot run x86_64 binaries
Reviewed by David Kilzer.

The check for whether WTF_USE_PLUGIN_HOST_PROCESS is defined occurs under the native architecture of the build machine.
If that is 32-bit then WTF_USE_PLUGIN_HOST_PROCESS will not be defined.  We work around this by forcing the check to
be performed against the x86_64 architecture.

* DerivedSources.make:

Identifier: 42946@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@51519 268f45cc-cd09-0410-ab3c-d52691b4dbfc