Fixes <http://webkit.org/b/31738>. Web Inspector: Console Scope Bar should be on top of console when Console is a Panel.
Reviewed by Pavel Feldman.

When the console is set to be shown as a full panel, move the scope bar to the top
where it looks best. When it is only shown as a drawer, move it to the bottom, where
it originally was. This takes up less space as a drawer, and makes it more discoverable
and usable when it is the full panel.

* inspector/front-end/ConsolePanel.js:
(WebInspector.ConsolePanel.prototype.show): Move filter bar to top.
(WebInspector.ConsolePanel.prototype.hide): Move filter bar back to bottom.
* inspector/front-end/inspector.css:



Identifier: 43377@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@51952 268f45cc-cd09-0410-ab3c-d52691b4dbfc