<rdar://problem/7512726> Flash full-screen does not work correctly with Flash 10.1 beta
Reviewed by Darin Adler.

Explicitly track when the plug-in shows and hides a full-screen window to determine
whether a plug-in is currently full-screen, rather than relying on menu bar visibility.
This ensures that we activate the plug-in host process and client application as
appropriate when the plug-in enters or exits full-screen.

* Plugins/Hosted/NetscapePluginHostProxy.h:
(WebKit::NetscapePluginHostProxy::isFullScreenWindowShowing):
* Plugins/Hosted/NetscapePluginHostProxy.mm:
(WebKit::NetscapePluginHostProxy::NetscapePluginHostProxy):
(WebKit::NetscapePluginHostProxy::setMenuBarVisible):
(WebKit::NetscapePluginHostProxy::didEnterFullScreen):
(WebKit::NetscapePluginHostProxy::didExitFullScreen):
(WebKit::NetscapePluginHostProxy::setFullScreenWindowIsShowing):
(WKPCSetMenuBarVisible):
(WKPCSetFullScreenWindowIsShowing):
* Plugins/Hosted/WebKitPluginClient.defs:

Identifier: 44586@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@53178 268f45cc-cd09-0410-ab3c-d52691b4dbfc