<rdar://problem/7573493> Error with line break inside ?&raquo; pair of characters. https://bugs.webkit.org/show_bug.cgi?id=17475
Reviewed by Darin Adler.

WebCore:

Test: fast/text/line-break-after-question-mark.html

Instead of unconditionally allowing lines to break after a question mark, which was intended
to mimic Internet Explorer, apply the Unicode line breaking behavior after a question mark,
with one exception in the ASCII range to match IE, namely not allowing a line break between
a question mark and a vertical line.

* rendering/break_lines.cpp:
(WebCore::shouldBreakAfter): Added a next character parameter. Changed to consult a table
for the question mark case in order to keep it fast and not require a text break iterator in
the ASCII case.
(WebCore::nextBreakablePosition): Pass the next character to shouldBreakAfter.

LayoutTests:

* fast/text/line-break-after-question-mark-expected.txt: Added.
* fast/text/line-break-after-question-mark.html: Added.
* fast/text/script-tests/line-break-after-question-mark.js: Added.
():
* platform/mac/tables/mozilla/bugs/bug6674-expected.checksum: Updated.
* platform/mac/tables/mozilla/bugs/bug6674-expected.png: Updated.
* platform/mac/tables/mozilla/bugs/bug6674-expected.txt: Updated.



Identifier: 45222@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@53839 268f45cc-cd09-0410-ab3c-d52691b4dbfc