WebCore: Drag and Drop: Windows uses "stop" sign as cursor when dragging https://bugs.webkit.org/show_bug.cgi?id=34305 <rdar://problem/7589672>
Reviewed by Adam Roben.

Add a FIXME for the code that needs to be changed to support full
custom cursors.

* page/EventHandler.cpp:
(WebCore::EventHandler::handleDrag):

WebKit/win: Drag and Drop: Windows uses "stop" sign as cursor when dragging
https://bugs.webkit.org/show_bug.cgi?id=34305
<rdar://problem/7589672>

Reviewed by Adam Roben.

Add a preference in WebKit (that defaults to false), for whether or not
we should show the custom cursors during drag and drop. However, this is
currently only used on Windows, and only used to hide the "drop not allowed"
icon inside the WebView is the preference is set to true.

This will be off by default, so no change in behavior.

* Interfaces/IWebPreferencesPrivate.idl: Added new functions.
* Interfaces/WebKit.idl: Touched to force Interfaces build.
* WebDropSource.cpp:
(WebDropSource::GiveFeedback): Implementation of conditional showing cursor
    logic.
* WebPreferenceKeysPrivate.h: Added new preference key.
* WebPreferences.cpp: Added new functions.
(WebPreferences::setCustomDragCursorsEnabled):
(WebPreferences::customDragCursorsEnabled):
* WebPreferences.h: Added new functions.



Identifier: 45450@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@54083 268f45cc-cd09-0410-ab3c-d52691b4dbfc