[Qt] WebSockets : Buffer the data in WebKit instead of QtNetwork https://bugs.webkit.org/show_bug.cgi?id=34425
Reviewed by Kenneth Rohde Christiansen.

Buffer the sent data in SocketStreamHandlePrivate instead of relying on
the network layer to do it. This is more robust and more consistent with how
Qt's HTTP stack works.
Close the socket in SocketStreamHandlePrivate::close() regardless of its state.

No new tests, since no new functionality is introduced.

* platform/network/qt/SocketStreamHandlePrivate.h:
* platform/network/qt/SocketStreamHandleQt.cpp:


Identifier: 45622@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@54279 268f45cc-cd09-0410-ab3c-d52691b4dbfc