[Gtk] Check if the renderer() exists in textForObject.
This is fixing a crash in the textForObject method. It can
happen when inspecting AtkObjects in the accerciser, then
navigating to a new page and still inspecting the old tree.

In this case the AccessibilityObjectWrapperAtk was still
valid but the included renderer() was returning 0. Add a
check for the accObject->renderer() before trying to use
the result.

* accessibility/gtk/AccessibilityObjectWrapperAtk.cpp:
(textForObject):

Identifier: 45673@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@54337 268f45cc-cd09-0410-ab3c-d52691b4dbfc