Removed some mistaken code added in http://trac.webkit.org/changeset/53860.
Reviewed by Oliver Hunt.

* API/APIShims.h:
(JSC::APICallbackShim::APICallbackShim):
(JSC::APICallbackShim::~APICallbackShim): No need to start/stop the
timeout checker when calling out from the API to the client; we want to
monitor the VM for timeouts, not the client. This mistake was harmless /
undetectable, since it's totally redundant with the APIEntryShim, which
also starts / stops the timeout checker.



Identifier: 46097@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@54790 268f45cc-cd09-0410-ab3c-d52691b4dbfc