[Qt] Avoid calling QFont::detach too often from FontQt
https://bugs.webkit.org/show_bug.cgi?id=35569

The letter- and wordspacing is inside the WebCore::Font class,
our QFont is located in the WebCore::FontPlatformData. Everytime
we need to use a QFont inside WebCore::Font we are calling the
font method which gets the QFont from the WebCore::FontPlatformData
and is applying the letter- and wordspacing. Internally this
will attempt to detach the QFont...

Avoid calling setLetterSpacing and setWordSpacing on QFont if
the WebCore::Font has the default applied.

* platform/graphics/qt/FontQt.cpp:
(WebCore::Font::font):

Identifier: 46815@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@55518 268f45cc-cd09-0410-ab3c-d52691b4dbfc