Make Document::postTask to use a single queue of tasks, to fire them in order https://bugs.webkit.org/show_bug.cgi?id=35943
Reviewed by Darin Adler.

JavaScriptCore:

The patch uses CFRunLoopTimer to schedule execution of tasks instead of performSelectorOnMainThread which apparently can starve other event sources.
The timer is used when the schedule request is coming on the main thread itself. This happens when the task is posted on the main thread or
when too many tasks are posted and the queue does 'stop and re-schedule' to make sure run loop has a chance to execute other events.

* wtf/mac/MainThreadMac.mm:
(WTF::timerFired):
(WTF::postTimer):
(WTF::scheduleDispatchFunctionsOnMainThread): Use timer posted to the current RunLoop if scheduling the task execution while on the main thread.

WebCore:

Test: existing worker-cloneport.html which was broken by initial patch in http://trac.webkit.org/changeset/55593.
Additional test which indirectly verifies the order of execution will come as part of https://bugs.webkit.org/show_bug.cgi?id=34726

* dom/Document.cpp:
(WebCore::Document::postTask): Always use the same task queue, independent of what thread is posting the task.

LayoutTests:

Updated the test since the order of events coming from independent task sources has changed.
This is a behavior change, but is consistent with the spec (which specifically says these events may be fired
in any order) and the last version of FF (3.5.8) which was completing the test successfully.

* http/tests/appcache/top-frame-2-expected.txt:
* http/tests/appcache/top-frame-2.html:

Identifier: 47111@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@55816 268f45cc-cd09-0410-ab3c-d52691b4dbfc