Fix for https://bugs.webkit.org/show_bug.cgi?id=33739 Fullscreen video HUD stays on top when switching to another window (e.g. via Alt-Tab) -and corresponding- <rdar://problem/7547574>
Reviewed by Adam Roben.

The HUD was always on top because it had the WS_EX_TOPMOST style.
So I removed the style and made m_videoWindow the owner of
m_hudWindow. This keeps m_hudWindow on top only when m_videoWindow
is the focused window.

* FullscreenVideoController.cpp:
(FullscreenVideoController::exitFullscreen): ASSERT that movie()->exitFullscreen() also destroyed the hud.
(FullscreenVideoController::createHUDWindow):



Identifier: 47237@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@55942 268f45cc-cd09-0410-ab3c-d52691b4dbfc